requirements.txt
PyYAML
tensorboard
torch
torchvision
tqdm
packaging
timm
torcheval
pillow
numpy
omegaconf
opencv-python
hydra-core
scikit-learn